Allow user delete only his pins

+++ pinry/core/static/core/js/pinry.js
@@ -79,9 +79,11 @@ function onLoadData(data) {
         html += '<div class="pin">';
         html += '   <div class="pin-options">';
 
-        html += '       <a href="/pins/delete-pin/'+image.id+'/">';
-        html += '           <div class="icon-trash"></div>';
-        html += '       </a>';
+        if (image.is_owner) {
+            html += '       <a href="/pins/delete-pin/'+image.id+'/">';
+            html += '           <div class="icon-trash"></div>';
+            html += '       </a>';
+        }
 
         html += '       <a href="/pins/'+image.id+'/like">';
         html += '           <div class="icon-heart"></div>';
